Make Enzyme tests use kernel wrappers from `src/ocn/Operators.jl`
  - Also moved the shared unit test definition code to common file
    that both Operator and Enzyme test can use.
