+++ test/enzyme/test_Enzyme_Operators.jl
@@ -1,11 +1,8 @@
 using Test
 using CUDA
+using MOKA
 using UnPack
 using CUDA: @allowscalar
-using MOKA: HorzMesh, ReadHorzMesh, GradientOnEdgeModified,
-            DivergenceOnCellModified1, DivergenceOnCellModified2,
-            Edge, Cell
-
 using Enzyme
 
 # Setting meshes to inactive types:
@@ -15,166 +12,9 @@ import Adapt
 import Downloads
 import KernelAbstractions as KA
 
-abstract type TestCase end 
-abstract type PlanarTest <: TestCase end 
+# include the testcase definition utilities
+include("../utilities.jl")
 
-atol = 1e-8
-
-on_architecture(backend::KA.Backend, array::AbstractArray) = Adapt.adapt_storage(backend, array)
-
-struct TestSetup{FT, AT}
-    
-    backend::KA.Backend
-
-    xᶜ::AT 
-    yᶜ::AT 
-
-    xᵉ::AT
-    yᵉ::AT
-
-    Lx::FT 
-    Ly::FT
-
-    EdgeNormalX::AT
-    EdgeNormalY::AT
-    
-    #TestSetup{FT,AT}(xᶜ, yᶜ, xᵉ, yᵉ, Lx, Ly, EdgeNormalX, EdgeNormalY) where {FT} = 
-    #    new{FT}(xᶜ, yᶜ, xᵉ, yᵉ, Lx, Ly, EdgeNormalX, EdgeNormalY)
-end 
-
-function TestSetup(mesh::HorzMesh, ::Type{PlanarTest}; backend=KA.CPU())
-
-    @unpack PrimaryCells, Edges = mesh
-
-    @unpack xᶜ, yᶜ = PrimaryCells 
-    @unpack xᵉ, yᵉ, angleEdge = Edges
-
-    FT = eltype(xᶜ)
-
-    #Lx = maximum(xᶜ) - minimum(xᶜ)
-    #Ly = maximum(yᶜ) - minimum(yᶜ)
-    Lx = round(maximum(xᶜ))
-    Ly = sqrt(3.0)/2.0 * Lx
-
-    EdgeNormalX = cos.(angleEdge)
-    EdgeNormalY = sin.(angleEdge)
-
-    return TestSetup(backend, 
-                     on_architecture(backend, xᶜ),
-                     on_architecture(backend, yᶜ),
-                     on_architecture(backend, xᵉ),
-                     on_architecture(backend, yᵉ), 
-                     Lx, Ly,
-                     on_architecture(backend, EdgeNormalX),
-                     on_architecture(backend, EdgeNormalY))
-end 
-
-"""
-Analytical function (defined as cell centers) 
-"""
-function h(test::TestSetup, ::Type{PlanarTest})
-        
-    @unpack xᶜ, yᶜ, Lx, Ly = test 
-
-    nCells = length(xᶜ)
-    ftype = eltype(xᶜ)
-    backend = KA.get_backend(xᶜ)
-    
-    result = KA.zeros(backend, ftype, (1, nCells))
-    result[1,:] = @. sin(2.0 * pi * xᶜ / Lx) * sin(2.0 * pi * yᶜ / Ly)
-
-    return result
-end
-
-"""
-"""
-function 𝐅ˣ(test::TestSetup, ::Type{PlanarTest})
-    @unpack xᵉ, yᵉ, Lx, Ly = test 
-
-    return @. sin(2.0 * pi * xᵉ / Lx) * cos(2.0 * pi * yᵉ / Ly)
-end
-
-"""
-"""
-function 𝐅ʸ(test::TestSetup, ::Type{PlanarTest})
-    @unpack xᵉ, yᵉ, Lx, Ly = test 
-
-    return @. cos(2.0 * pi * xᵉ / Lx) * sin(2.0 * pi * yᵉ / Ly)
-end
-
-"""
-The edge normal component of the vector field of 𝐅
-"""
-function 𝐅ₑ(test::TestSetup, ::Type{TC}) where {TC <: TestCase} 
-
-    @unpack EdgeNormalX, EdgeNormalY = test
-
-    # need intermediate values from broadcasting to work correctly
-    𝐅ˣᵢ = 𝐅ˣ(test, TC)
-    𝐅ʸᵢ = 𝐅ʸ(test, TC)
-    
-    ftype = eltype(EdgeNormalX)
-    nEdges = length(EdgeNormalX)
-    backend = KA.get_backend(EdgeNormalX)
-    
-    result = KA.zeros(backend, ftype, (1, nEdges))
-
-    result[1,:] = @. EdgeNormalX * 𝐅ˣᵢ + EdgeNormalY * 𝐅ʸᵢ
-
-    return result
-end
-
-function gradient!(grad, hᵢ, mesh::HorzMesh; backend=KA.CPU())
-    
-    @unpack Edges = mesh
-
-    @unpack nEdges, dcEdge, cellsOnEdge = Edges
-    
-    # only testing horizontal mesh, so set up dummy array for verticalLevels
-    vert_levels = 1
-
-    # New modified kernel:
-    kernel! = GradientOnEdgeModified(backend)
-    kernel!(cellsOnEdge, dcEdge, hᵢ, grad, workgroupsize=64, ndrange=(nEdges, vert_levels))
-
-    KA.synchronize(backend)
-end
-
-function divergence!(div, 𝐅ₑ, mesh::HorzMesh; backend=KA.CPU())
-
-    @unpack PrimaryCells, DualCells, Edges = mesh
-
-    @unpack nEdges, dvEdge = Edges
-    @unpack nCells, nEdgesOnCell = PrimaryCells
-    @unpack edgesOnCell, edgeSignOnCell, areaCell = PrimaryCells
-
-    # only testing horizontal mesh, so set up dummy array for verticalLevels
-    vert_levels = 1
-    
-    kernel1! = DivergenceOnCellModified1(backend)
-    kernel2! = DivergenceOnCellModified2(backend)
-    
-    kernel1!(𝐅ₑ, dvEdge, workgroupsize=64, ndrange=(nEdges, vert_levels))
-
-    kernel2!(div,
-            𝐅ₑ,
-            nEdgesOnCell,
-            edgesOnCell,
-            edgeSignOnCell,
-            areaCell,
-            workgroupsize=32,
-            ndrange=(nCells, vert_levels))
-    
-    KA.synchronize(backend)
-end
-
-# NOTE: planar doubly periodic meshes on lcrc do not give the expected answers
-#       following Omega devGuide and using a custom generated mesh
-#lcrc_url="https://web.lcrc.anl.gov/public/e3sm/mpas_standalonedata/mpas-ocean/"
-#mesh_fp ="mesh_database/doubly_periodic_20km_1000x2000km_planar.151027.nc"
-#mesh_fp ="mesh_database/doubly_periodic_10km_1000x2000km_planar.151117.nc"
-#mesh_url = lcrc_url * mesh_fp
-#
 mesh_url = "https://gist.github.com/mwarusz/f8caf260398dbe140d2102ec46a41268/raw/e3c29afbadc835797604369114321d93fd69886d/PlanarPeriodic48x48.nc"
 mesh_fn  = "MokaMesh.nc"
 
@@ -183,17 +23,26 @@ Downloads.download(mesh_url, mesh_fn)
 backend = KA.CPU()
 #backend = CUDABackend();
 
-mesh = ReadHorzMesh(mesh_fn; backend=backend)
-setup = TestSetup(mesh, PlanarTest; backend=backend)
+# Read in the purely horizontal doubly periodic testing mesh
+HorzMesh = ReadHorzMesh(mesh_fn; backend=backend)
+# Create a dummy vertical mesh from the horizontal mesh
+VertMesh = VerticalMesh(HorzMesh; nVertLevels=1, backend=backend)
+# Create a the full Mesh strucutre 
+MPASMesh = Mesh(HorzMesh, VertMesh)
 
+setup = TestSetup(MPASMesh, PlanarTest; backend=backend)
+
+nEdges = HorzMesh.Edges.nEdges
+nCells = HorzMesh.PrimaryCells.nCells
+nVertLevels = VertMesh.nVertLevels
 
 ###
 ### Here, we will test Enzyme AD on our kernels
 ###
 
 # As a clean / easy to read test, let's create an outer function that measures the squared norm of the gradient computed by kernel:
-function gradient_normSq(grad, hᵢ, mesh::HorzMesh; backend=KA.CPU())
-    gradient!(grad, hᵢ, mesh::HorzMesh; backend=backend)
+function gradient_normSq(grad, hᵢ, mesh::Mesh; backend=KA.CPU())
+    GradientOnEdge!(grad, hᵢ, mesh::Mesh; backend=backend)
 
     normSq = 0.0
     N = size(grad)
@@ -205,24 +54,24 @@ function gradient_normSq(grad, hᵢ, mesh::HorzMesh; backend=KA.CPU())
 end
 
 # Let's recreate all the variables:
-gradNum = KA.zeros(backend, Float64, (1, mesh.Edges.nEdges))
+gradNum = KA.zeros(backend, Float64, (nVertLevels, nEdges))
 Scalar  = h(setup, PlanarTest)
 
-d_gradNum = KA.zeros(backend, Float64, (1, mesh.Edges.nEdges))
-d_Scalar  = KA.zeros(backend, eltype(setup.xᶜ), (1, size(setup.xᶜ)[1]))
-d_mesh    = Enzyme.make_zero(mesh)
+d_gradNum  = KA.zeros(backend, Float64, (nVertLevels, nEdges))
+d_Scalar   = KA.zeros(backend, eltype(setup.xᶜ), (nVertLevels, nCells))
+d_MPASMesh = Enzyme.make_zero(MPASMesh)
 
 d_normSq = autodiff(Enzyme.Reverse,
                     gradient_normSq,
                     Duplicated(gradNum, d_gradNum),
                     Duplicated(Scalar, d_Scalar),
-                    Duplicated(mesh, d_mesh))
+                    Duplicated(MPASMesh, d_MPASMesh))
 
 #@show d_gradNum
 #@show d_Scalar
 
 # For comparison, let's compute the derivative by hand for a given scalar entry:
-gradNum = KA.zeros(backend, Float64, (1, mesh.Edges.nEdges))
+gradNum = KA.zeros(backend, Float64, (nVertLevels, nEdges))
 Scalar  = h(setup, PlanarTest)
 ScalarP = deepcopy(Scalar)
 ScalarM = deepcopy(Scalar)
@@ -232,9 +81,9 @@ k = 1837
 ScalarP[k] = ScalarP[k] + abs(ScalarP[k]) * ϵ
 ScalarM[k] = ScalarM[k] - abs(ScalarM[k]) * ϵ
 
-normSqP = gradient_normSq(gradNum, ScalarP, mesh)
-gradNum = KA.zeros(backend, Float64, (1, mesh.Edges.nEdges))
-normSqM = gradient_normSq(gradNum, ScalarM, mesh)
+normSqP = gradient_normSq(gradNum, ScalarP, MPASMesh)
+gradNum = KA.zeros(backend, Float64, (nVertLevels, nEdges))
+normSqM = gradient_normSq(gradNum, ScalarM, MPASMesh)
 
 dnorm_dscalar_fd = (normSqP - normSqM) / (ScalarP[k] - ScalarM[k])
 dnorm_dscalar    = d_Scalar[k]
@@ -244,12 +93,11 @@ For edge global index $k
 Enzyme computed $dnorm_dscalar
 Finite differences computed $dnorm_dscalar_fd
 """
-
 ###
 ### Now let's test divergence:
 ###
-function divergence_normSq(div, 𝐅ₑ, mesh::HorzMesh; backend=KA.CPU())
-    divergence!(div, 𝐅ₑ, mesh::HorzMesh; backend=backend)
+function divergence_normSq(div, 𝐅ₑ, mesh::Mesh; backend=KA.CPU())
+    DivergenceOnCell!(div, 𝐅ₑ, mesh::Mesh; backend=backend)
 
     normSq = 0.0
     N = size(div)
@@ -260,17 +108,17 @@ function divergence_normSq(div, 𝐅ₑ, mesh::HorzMesh; backend=KA.CPU())
     return normSq
 end
 
-divNum  = KA.zeros(backend, Float64, (1, mesh.PrimaryCells.nCells))
+divNum  = KA.zeros(backend, Float64, (nVertLevels, nCells))
 VecEdge = 𝐅ₑ(setup, PlanarTest)
 
-d_divNum  = KA.zeros(backend, Float64, (1, mesh.PrimaryCells.nCells))
-d_VecEdge = KA.zeros(backend, eltype(setup.EdgeNormalX), (1, size(setup.EdgeNormalX)[1]))
+d_divNum  = KA.zeros(backend, Float64, (nVertLevels, nCells))
+d_VecEdge = KA.zeros(backend, eltype(setup.EdgeNormalX), (nVertLevels, nEdges))
 
 d_normSq = autodiff(Enzyme.Reverse,
                     divergence_normSq,
                     Duplicated(divNum, d_divNum),
                     Duplicated(VecEdge, d_VecEdge),
-                    Duplicated(mesh, d_mesh))
+                    Duplicated(MPASMesh, d_MPASMesh))
 
 # For comparison, let's compute the derivative by hand for a given VecEdge entry:
 VecEdgeP = 𝐅ₑ(setup, PlanarTest)
@@ -284,10 +132,10 @@ VecEdgeM[k] = VecEdgeM[k] - abs(VecEdgeM[k]) * ϵ
 VecEdgePk = VecEdgeP[k]
 VecEdgeMk = VecEdgeM[k]
 
-divNum  = KA.zeros(backend, Float64, (1, mesh.PrimaryCells.nCells))
-normSqP = divergence_normSq(divNum, VecEdgeP, mesh)
-divNum  = KA.zeros(backend, Float64, (1, mesh.PrimaryCells.nCells))
-normSqM = divergence_normSq(divNum, VecEdgeM, mesh)
+divNum  = KA.zeros(backend, Float64, (nVertLevels, nCells))
+normSqP = divergence_normSq(divNum, VecEdgeP, MPASMesh)
+divNum  = KA.zeros(backend, Float64, (nVertLevels, nCells))
+normSqM = divergence_normSq(divNum, VecEdgeM, MPASMesh)
 
 dnorm_dvecedge_fd = (normSqP - normSqM) / (VecEdgePk - VecEdgeMk)
 dnorm_dvecedge    = d_VecEdge[k]
